Move old implementation to legacy folder
