+++ client/examples/saturate.rs
@@ -18,8 +18,8 @@ use tokio::io::AsyncWriteExt;
 use tokio::sync::RwLock;
 use tokio_stream::wrappers::UnboundedReceiverStream;
 
-use ya_relay_client::client::ForwardSender;
-use ya_relay_client::{ClientBuilder, ForwardReceiver};
+use ya_relay_client::legacy::client::ForwardSender;
+use ya_relay_client::legacy::{ClientBuilder, ForwardReceiver};
 use ya_relay_core::crypto::FallbackCryptoProvider;
 use ya_relay_core::key::{load_or_generate, Protected};
 use ya_relay_core::server_session::TransportType;
