+++ client/src/legacy.rs
@@ -1,0 +1,9 @@
+pub mod client;
+mod expire;
+mod registry;
+mod session_guard;
+mod session_manager;
+mod session_start;
+mod virtual_layer;
+
+pub use client::{Client, ClientBuilder, ForwardReceiver};
