+++ client/src/legacy/registry.rs
