+++ client/src/legacy/session_guard.rs
