+++ client/src/testing.rs
@@ -1,7 +1,7 @@
 pub mod _forwarding_utils;
 pub mod accessors;
-pub mod forwarding_utils;
 pub mod init;
+pub mod legacy_forwarding_utils;
 pub mod mocks;
 
 pub use ya_relay_core::session::Session;
