+++ client/tests/test_forwarding.rs
@@ -9,8 +9,8 @@ use tokio::time::timeout;
 use tokio_stream::wrappers::UnboundedReceiverStream;
 
 use ya_relay_client::_client::{Forwarded, GenericSender};
-use ya_relay_client::testing::forwarding_utils::spawn_receive;
 use ya_relay_client::testing::init::MockSessionNetwork;
+use ya_relay_client::testing::legacy_forwarding_utils::spawn_receive;
 use ya_relay_core::server_session::TransportType;
 
 #[serial_test::serial]
