+++ server/tests/test_closed_session.rs
@@ -2,7 +2,7 @@ use std::collections::{HashMap, HashSet};
 use std::hash::Hash;
 use std::time::Duration;
 use test_case::test_case;
-use ya_relay_client::ClientBuilder;
+use ya_relay_client::legacy::ClientBuilder;
 use ya_relay_core::crypto::{CryptoProvider, FallbackCryptoProvider};
 use ya_relay_core::key::generate;
 use ya_relay_server::testing::server::init_test_server;
