+++ server/tests/test_discovery.rs
@@ -7,8 +7,8 @@ use std::sync::atomic::AtomicBool;
 use std::sync::atomic::Ordering::SeqCst;
 use std::time::Duration;
 
-use ya_relay_client::testing::forwarding_utils::spawn_receive;
-use ya_relay_client::ClientBuilder;
+use ya_relay_client::legacy::ClientBuilder;
+use ya_relay_client::testing::legacy_forwarding_utils::spawn_receive;
 use ya_relay_core::crypto::{CryptoProvider, FallbackCryptoProvider};
 use ya_relay_core::key::generate;
 use ya_relay_server::testing::server::init_test_server;
