+++ server/tests/test_expired_session.rs
@@ -1,9 +1,9 @@
 use std::time::Duration;
 
-use ya_relay_client::testing::forwarding_utils::{
+use ya_relay_client::legacy::ClientBuilder;
+use ya_relay_client::testing::legacy_forwarding_utils::{
     check_broadcast, check_forwarding, spawn_receive_for_client, Mode,
 };
-use ya_relay_client::ClientBuilder;
 use ya_relay_core::utils::to_udp_url;
 use ya_relay_server::testing::server::{
     init_test_server, init_test_server_with_config, test_default_config,
